integration tests for addDeviceStart and updateDevice
